+++ src/main/java/com/authlete/jakarta/api/AuthleteApiImplV3.java
@@ -28,11 +28,11 @@ import com.authlete.common.api.AuthleteApiException;
 import com.authlete.common.conf.AuthleteApiVersion;
 import com.authlete.common.conf.AuthleteConfiguration;
 import com.authlete.common.dto.*;
+import com.authlete.common.types.TokenStatus;
 
 
 /**
  * The implementation of {@link AuthleteApi} using JAX-RS 2.0 client API.
- *
  * Supports Authlete API V3.
  *
  * @author Justin Richer
@@ -113,10 +113,13 @@ public class AuthleteApiImplV3 extends AuthleteApiJaxrsImpl
     private static final String VCI_DEFERRED_PARSE_API_PATH                   = "/api/%d/vci/deferred/parse";
     private static final String VCI_DEFERRED_ISSUE_API_PATH                   = "/api/%d/vci/deferred/issue";
     private static final String ID_TOKEN_REISSUE_API_PATH                     = "/api/%d/idtoken/reissue";
+    private static final String TOKEN_CREATE_BATCH_API_PATH                   = "/api/%d/token/create/batch";
+    private static final String TOKEN_CREATE_BATCH_STATUS_API_PATH            = "/api/%d/token/create/batch/status";
 
 
     private final String mAuth;
     private final Long mServiceId;
+
 
     /**
      * The constructor with an instance of {@link AuthleteConfiguration}.
@@ -416,48 +419,76 @@ public class AuthleteApiImplV3 extends AuthleteApiJaxrsImpl
     @Override
     public TokenListResponse getTokenList() throws AuthleteApiException
     {
-        return getTokenList(null, null, 0, 0, false);
+        return getTokenList(null, null, 0, 0, false, TokenStatus.ALL);
+    }
+
+
+    @Override
+    public TokenListResponse getTokenList(TokenStatus tokenStatus) throws AuthleteApiException
+    {
+        return getTokenList(null, null, 0, 0, false, tokenStatus);
     }
 
 
     @Override
     public TokenListResponse getTokenList(String clientIdentifier, String subject) throws AuthleteApiException
     {
-        return getTokenList(clientIdentifier, subject, 0, 0, false);
+        return getTokenList(clientIdentifier, subject, 0, 0, false, TokenStatus.ALL);
+    }
+
+
+    @Override
+    public TokenListResponse getTokenList(String clientIdentifier, String subject, TokenStatus tokenStatus) throws AuthleteApiException
+    {
+        return getTokenList(clientIdentifier, subject, 0, 0, false, tokenStatus);
     }
 
 
     @Override
     public TokenListResponse getTokenList(int start, int end) throws AuthleteApiException
     {
-        return getTokenList(null, null, start, end, true);
+        return getTokenList(null, null, start, end, true, TokenStatus.ALL);
+    }
+
+
+    @Override
+    public TokenListResponse getTokenList(int start, int end, TokenStatus tokenStatus) throws AuthleteApiException
+    {
+        return getTokenList(null, null, start, end, true, tokenStatus);
     }
 
 
     @Override
     public TokenListResponse getTokenList(String clientIdentifier, String subject, int start, int end) throws AuthleteApiException
     {
-        return getTokenList(clientIdentifier, subject, start, end, true);
+        return getTokenList(clientIdentifier, subject, start, end, true, TokenStatus.ALL);
+    }
+
+
+    @Override
+    public TokenListResponse getTokenList(String clientIdentifier, String subject, int start, int end, TokenStatus tokenStatus) throws AuthleteApiException
+    {
+        return getTokenList(clientIdentifier, subject, start, end, true, tokenStatus);
     }
 
 
     private TokenListResponse getTokenList(
             final String clientIdentifier, final String subject,
-            final int start, final int end, final boolean rangeGiven) throws AuthleteApiException
+            final int start, final int end, final boolean rangeGiven, TokenStatus tokenStatus) throws AuthleteApiException
     {
         return executeApiCall(new AuthleteApiCall<TokenListResponse>()
         {
             @Override
             public TokenListResponse call()
             {
-                return callGetTokenList(clientIdentifier, subject, start, end, rangeGiven);
+                return callGetTokenList(clientIdentifier, subject, start, end, rangeGiven, tokenStatus);
             }
         });
     }
 
 
     private TokenListResponse callGetTokenList(
-            String clientIdentifier, String subject, int start, int end, boolean rangeGiven)
+            String clientIdentifier, String subject, int start, int end, boolean rangeGiven, TokenStatus tokenStatus)
     {
         String path = String.format(AUTH_TOKEN_GET_LIST_API_PATH, mServiceId);
 
@@ -477,6 +508,8 @@ public class AuthleteApiImplV3 extends AuthleteApiJaxrsImpl
         {
             target = target.queryParam("start", start).queryParam("end", end);
         }
+
+        target = target.queryParam("tokenStatus", tokenStatus.toString());
 
         // FIXME: it feels weird that this is in its own space instead of the caller classes, is there a reason for that?
         return wrapWithDpop(target
@@ -1459,5 +1492,28 @@ public class AuthleteApiImplV3 extends AuthleteApiJaxrsImpl
                 new PostApiCaller<AuthorizationTicketUpdateResponse>(
                         AuthorizationTicketUpdateResponse.class, request,
                         AUTH_AUTHORIZATION_TICKET_UPDATE_API_PATH, mServiceId));
+    }
+
+
+    @Override
+    public TokenCreateBatchResponse tokenCreateBatch(
+            TokenCreateRequest[] request, boolean dryRun) throws AuthleteApiException
+    {
+        return executeApiCall(
+                new PostApiCaller<TokenCreateBatchResponse>(
+                        TokenCreateBatchResponse.class, request,
+                        TOKEN_CREATE_BATCH_API_PATH, mServiceId)
+                .addParam("dryRun", dryRun));
+    }
+
+
+    @Override
+    public TokenCreateBatchStatusResponse getTokenCreateBatchStatus(
+            String requestId) throws AuthleteApiException
+    {
+        return executeApiCall(
+                new GetApiCaller<TokenCreateBatchStatusResponse>(
+                        TokenCreateBatchStatusResponse.class,
+                        TOKEN_CREATE_BATCH_STATUS_API_PATH, mServiceId, requestId));
     }
 }